+++ DDL/users.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `users` (
-  `id` int NOT NULL AUTO_INCREMENT,
+  `id` bigint NOT NULL AUTO_INCREMENT,
   `name` varchar(32) NOT NULL,
   `external_service_id` varchar(50) NOT NULL,
   `external_service_type` varchar(16) NOT NULL,
